refactor(api): move chat oauth logic to intgr controller
