+++ apps/api/src/app/integrations/dtos/generate-chat-oauth-url.dto.ts
