+++ apps/api/src/app/integrations/integrations.module.ts
@@ -1,17 +1,29 @@
 import { forwardRef, Module } from '@nestjs/common';
-import { ChannelFactory, CompileTemplate, CreateExecutionDetails } from '@novu/application-generic';
-import { CommunityOrganizationRepository } from '@novu/dal';
+import {
+  CalculateLimitNovuIntegration,
+  ChannelFactory,
+  CompileTemplate,
+  CreateExecutionDetails,
+  GetNovuProviderCredentials,
+} from '@novu/application-generic';
+import { ChannelConnectionRepository, CommunityOrganizationRepository, CommunityUserRepository } from '@novu/dal';
 import { AuthModule } from '../auth/auth.module';
 import { SharedModule } from '../shared/shared.module';
 import { IntegrationsController } from './integrations.controller';
 import { USE_CASES } from './usecases';
 
-const PROVIDERS = [ChannelFactory, CompileTemplate];
+const PROVIDERS = [ChannelFactory, CompileTemplate, GetNovuProviderCredentials, CalculateLimitNovuIntegration];
 
 @Module({
   imports: [SharedModule, forwardRef(() => AuthModule)],
   controllers: [IntegrationsController],
-  providers: [...USE_CASES, CreateExecutionDetails, CommunityOrganizationRepository, ...PROVIDERS],
+  providers: [
+    ...USE_CASES,
+    CommunityOrganizationRepository,
+    CommunityUserRepository,
+    ChannelConnectionRepository,
+    ...PROVIDERS,
+  ],
   exports: [...USE_CASES],
 })
 export class IntegrationModule {}
