+++ apps/api/src/app/integrations/usecases/index.ts
@@ -5,11 +5,16 @@ import {
   NormalizeVariables,
   SelectIntegration,
 } from '@novu/application-generic';
+import { CreateChannelConnection } from '../../channel-connections/usecases/create-channel-connection/create-channel-connection.usecase';
 import { AutoConfigureIntegration } from './auto-configure-integration/auto-configure-integration.usecase';
+import { ChatOauthCallback } from './chat-oauth-callback/chat-oauth-callback.usecase';
+import { SlackOauthCallback } from './chat-oauth-callback/slack-oauth-callback/slack-oauth-callback.usecase';
 import { CheckIntegration } from './check-integration/check-integration.usecase';
 import { CheckIntegrationEMail } from './check-integration/check-integration-email.usecase';
 import { CreateIntegration } from './create-integration/create-integration.usecase';
 import { CreateNovuIntegrations } from './create-novu-integrations/create-novu-integrations.usecase';
+import { GenerateChatOauthUrl } from './generate-chat-oath-url/generate-chat-oauth-url.usecase';
+import { GenerateSlackOauthUrl } from './generate-chat-oath-url/generate-slack-oath-url/generate-slack-oauth-url.usecase';
 import { GetActiveIntegrations } from './get-active-integration/get-active-integration.usecase';
 import { GetInAppActivated } from './get-in-app-activated/get-in-app-activated.usecase';
 import { GetIntegrations } from './get-integrations/get-integrations.usecase';
@@ -36,4 +41,9 @@ export const USE_CASES = [
   SetIntegrationAsPrimary,
   CreateNovuIntegrations,
   NormalizeVariables,
+  GenerateChatOauthUrl,
+  GenerateSlackOauthUrl,
+  SlackOauthCallback,
+  ChatOauthCallback,
+  CreateChannelConnection,
 ];
