+++ apps/api/src/app/subscribers-v2/subscribers.controller.ts
@@ -1,16 +1,13 @@
 import {
-  BadRequestException,
   Body,
   ClassSerializerInterceptor,
   Controller,
   Delete,
   Get,
-  NotFoundException,
   Param,
   Patch,
   Post,
   Query,
-  Res,
   UseInterceptors,
 } from '@nestjs/common';
 import { ApiExcludeEndpoint, ApiOperation, ApiParam, ApiQuery, ApiTags } from '@nestjs/swagger';
@@ -18,19 +15,16 @@ import {
   CreateOrUpdateSubscriberCommand,
   CreateOrUpdateSubscriberUseCase,
   ExternalApiAccessible,
-  FeatureFlagsService,
   RequirePermissions,
   UserSession,
 } from '@novu/application-generic';
 import {
   ApiRateLimitCategoryEnum,
   DirectionEnum,
-  FeatureFlagsKeysEnum,
   PermissionsEnum,
   SubscriberCustomData,
   UserSessionData,
 } from '@novu/shared';
-import { Response } from 'express';
 import { RequireAuthentication } from '../auth/framework/auth.decorator';
 import { GetPreferencesResponseDto } from '../inbox/dtos/get-preferences-response.dto';
 import { BulkUpdatePreferencesCommand } from '../inbox/usecases/bulk-update-preferences/bulk-update-preferences.command';
@@ -49,7 +43,6 @@ import { ListSubscriberSubscriptionsCommand } from '../topics-v2/usecases/list-s
 import { ListSubscriberSubscriptionsUseCase } from '../topics-v2/usecases/list-subscriber-subscriptions/list-subscriber-subscriptions.usecase';
 import { BulkUpdateSubscriberPreferencesDto } from './dtos/bulk-update-subscriber-preferences.dto';
 import { CreateSubscriberRequestDto } from './dtos/create-subscriber.dto';
-import { GenerateChatOauthUrlRequestDto } from './dtos/generate-chat-oauth-url.dto';
 import { GetSubscriberPreferencesDto } from './dtos/get-subscriber-preferences.dto';
 import { GetSubscriberPreferencesRequestDto } from './dtos/get-subscriber-preferences-request.dto';
 import { ListSubscribersQueryDto } from './dtos/list-subscribers-query.dto';
@@ -58,11 +51,6 @@ import { PatchSubscriberRequestDto } from './dtos/patch-subscriber.dto';
 import { PatchSubscriberPreferencesDto } from './dtos/patch-subscriber-preferences.dto';
 import { RemoveSubscriberResponseDto } from './dtos/remove-subscriber.dto';
 import { SubscriberGlobalPreferenceDto } from './dtos/subscriber-global-preference.dto';
-import { ChatOauthCallbackCommand } from './usecases/chat-oauth-callback/chat-oauth-callback.command';
-import { ResponseTypeEnum } from './usecases/chat-oauth-callback/chat-oauth-callback.response';
-import { ChatOauthCallback } from './usecases/chat-oauth-callback/chat-oauth-callback.usecase';
-import { GenerateChatOauthUrlCommand } from './usecases/generate-chat-oath-url/generate-chat-oauth-url.command';
-import { GenerateChatOauthUrl } from './usecases/generate-chat-oath-url/generate-chat-oauth-url.usecase';
 import { GetSubscriberCommand } from './usecases/get-subscriber/get-subscriber.command';
 import { GetSubscriber } from './usecases/get-subscriber/get-subscriber.usecase';
 import { GetSubscriberPreferencesCommand } from './usecases/get-subscriber-preferences/get-subscriber-preferences.command';
@@ -94,9 +82,6 @@ export class SubscribersController {
     private bulkUpdatePreferencesUsecase: BulkUpdatePreferences,
     private createOrUpdateSubscriberUsecase: CreateOrUpdateSubscriberUseCase,
     private listSubscriberSubscriptionsUsecase: ListSubscriberSubscriptionsUseCase,
-    private chatOauthCallbackUsecase: ChatOauthCallback,
-    private generateChatOauthUrlUsecase: GenerateChatOauthUrl,
-    private featureFlagsService: FeatureFlagsService,
     private getSubscriberGlobalPreference: GetSubscriberGlobalPreference
   ) {}
 
@@ -410,84 +395,5 @@ export class SubscribersController {
         includeCursor: query.includeCursor,
       })
     );
-  }
-
-  @Post('/chat/oauth')
-  @ApiOperation({
-    summary: 'Generate chat OAuth URL',
-    description: `Generate an OAuth URL for chat integrations like Slack. 
-    The subscriber will use this URL to authorize the chat integration.`,
-  })
-  @ApiResponse(String)
-  @ApiExcludeEndpoint()
-  @RequirePermissions(PermissionsEnum.SUBSCRIBER_WRITE)
-  @RequireAuthentication()
-  async getChatOAuthUrl(
-    @UserSession() user: UserSessionData,
-    @Body() body: GenerateChatOauthUrlRequestDto
-  ): Promise<string> {
-    await this.checkFeatureEnabled(user);
-
-    return await this.generateChatOauthUrlUsecase.execute(
-      GenerateChatOauthUrlCommand.create({
-        environmentId: user.environmentId,
-        organizationId: user.organizationId,
-        subscriberId: body.subscriberId,
-        integrationIdentifier: body.integrationIdentifier,
-        providerId: body.providerId,
-      })
-    );
-  }
-
-  @Get('/chat/oauth/callback')
-  @ApiOperation({
-    summary: 'Handle chat OAuth callback',
-    description: `Generic OAuth callback handler for all chat integrations (Slack, Teams, Discord, etc.). 
-    This endpoint processes the authorization code and stores the connection for any supported chat provider.`,
-  })
-  @ApiExcludeEndpoint()
-  async handleChatOAuthCallback(
-    @Res() res: Response,
-    @Query('code') providerCode: string,
-    @Query('state') state: string,
-    @Query('error') error?: string,
-    @Query('error_description') errorDescription?: string
-  ): Promise<void> {
-    if (error) {
-      throw new BadRequestException(`OAuth error: ${error}${errorDescription ? ` - ${errorDescription}` : ''}`);
-    }
-
-    if (!providerCode || !state) {
-      throw new BadRequestException('Missing required OAuth parameters: code and state');
-    }
-
-    const result = await this.chatOauthCallbackUsecase.execute(
-      ChatOauthCallbackCommand.create({
-        providerCode,
-        state,
-      })
-    );
-
-    if (result.type === ResponseTypeEnum.HTML) {
-      res.setHeader('Content-Type', 'text/html');
-      res.setHeader('Content-Security-Policy', "default-src 'self'; script-src 'self' 'unsafe-inline'");
-      res.send(result.result);
-
-      return;
-    }
-
-    res.redirect(result.result);
-  }
-
-  private async checkFeatureEnabled(user: UserSessionData) {
-    const isEnabled = await this.featureFlagsService.getFlag({
-      key: FeatureFlagsKeysEnum.IS_SLACK_TEAMS_ENABLED,
-      defaultValue: false,
-      organization: { _id: user.organizationId },
-    });
-
-    if (!isEnabled) {
-      throw new NotFoundException('Feature not enabled');
-    }
   }
 }
